% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataDict.R
\name{print_data_dict}
\alias{print_data_dict}
\title{Print Data Dictionary}
\usage{
print_data_dict(data, title = "", descriptions = NULL)
}
\arguments{
\item{data}{The dataset for which the data dictionary is to be printed.}

\item{title}{The title for the data dictionary (optional).}

\item{descriptions}{A data frame providing detailed descriptions for
variables (optional).}
}
\value{
None
}
\description{
Creates a data dictionary for the dataset using attributes from the data
and description if provided.
Adapted from the data_dict_md function from the explore package.
Source: https://rdrr.io/cran/explore/src/R/tools.R
}
\examples{
print_data_dict(mtcars, title = "Cars", descriptions = descriptions)
## **Cars**
**Number of Rows**: `32`

|Variable |Type    | NA| \%NA| Unique|Description      |
  |:--------|:-------|--:|---:|------:|:----------------|
  |mpg      |numeric |  0|   0|     25|mpg_description  |
  |cyl      |numeric |  0|   0|      3|cyl_description  |
  |disp     |numeric |  0|   0|     27|disp_description |
  |hp       |numeric |  0|   0|     22|hp_description   |
  |drat     |numeric |  0|   0|     22|drat_description |
  |wt       |numeric |  0|   0|     29|wt_description   |
  |qsec     |numeric |  0|   0|     30|qsec_description |
  |vs       |numeric |  0|   0|      2|vs_description   |
  |am       |numeric |  0|   0|      2|am_description   |
  |gear     |numeric |  0|   0|      3|gear_description |
  |carb     |numeric |  0|   0|      6|carb_description |
}
