% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataDict.R
\name{print_descriptions_df}
\alias{print_descriptions_df}
\title{Generate Skeleton for Variable Descriptions}
\usage{
print_descriptions_df(data)
}
\arguments{
\item{data}{The dataset for which the data dictionary is to be printed.}
}
\value{
None
}
\description{
Generates a skeleton code to create a dataframe containing variable names
and their corresponding descriptions.
It is designed to simplify the process of creating input for
\code{\link{print_data_dict}} function by providing a template for describing
variables in a dataset.
}
\examples{
>print_descriptions_df(mtcars)

descriptions = data.frame(Variable = c('mpg',
'cyl',
'disp',
'hp',
'drat',
'wt',
'qsec',
'vs',
'am',
'gear',
'carb'),
Description = c('mpg_description',
                'cyl_description',
                'disp_description',
                'hp_description',
                'drat_description',
                'wt_description',
                'qsec_description',
                'vs_description',
                'am_description',
                'gear_description',
                'carb_description'))
}
